+++ src/components/HeaderBlock/style.module.css
@@ -1,5 +1,6 @@
 .root {
   background: #202736;
+  background: linear-gradient(to bottom, #181d23 0%, #202736 80%);
   background-attachment: fixed;
   background-size: cover;
   position: relative;
@@ -8,16 +9,16 @@
   z-index: 1;
 }
 
-.root::before {
+.root:before {
   content: "";
   position: fixed;
-  background: url(../../assets/Stars.svg);
+  background: url(https://raw.githubusercontent.com/yagoestevez/fcc-portfolio/master/src/Images/Stars.svg?sanitize=true);
   background-attachment: fixed;
   width: 100%;
   min-height: 100vh;
   z-index: -1;
-  animation: stars-move-in 1000ms 300ms forwards;
   opacity: 0;
+  animation: stars-move-in 1000ms 300ms forwards;
 }
 
 @keyframes stars-move-in {
@@ -34,7 +35,8 @@
   position: absolute;
   bottom: -300px;
   left: 0;
-  background: url(../../assets/Trees.svg) bottom left repeat-x;
+  background: url(https://raw.githubusercontent.com/yagoestevez/fcc-portfolio/master/src/Images/Trees.svg?sanitize=true)
+    bottom left repeat-x;
   background-size: cover;
   width: 100%;
   height: 80%;
@@ -53,12 +55,61 @@
   }
 }
 
+.silhouette {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  background: url(https://raw.githubusercontent.com/yagoestevez/fcc-portfolio/master/src/Images/Silhouette.svg?sanitize=true)
+    bottom left no-repeat;
+  width: 50%;
+  height: 50%;
+  opacity: 0;
+  animation: silhouette-move-in 1000ms 800ms forwards;
+}
+
+@keyframes silhouette-move-in {
+  from {
+    background-position-x: 0;
+  }
+  to {
+    opacity: 1;
+    background-position-x: 50%;
+  }
+}
+
+.moon {
+  position: absolute;
+  top: 0;
+  right: 0;
+  position: fixed;
+  background: url(https://raw.githubusercontent.com/yagoestevez/fcc-portfolio/master/src/Images/Moon.svg?sanitize=true)
+    right 150% no-repeat;
+  background-size: 30% 30%;
+  background-attachment: fixed;
+  width: 100%;
+  height: 100%;
+  z-index: -1;
+  opacity: 0;
+  animation: moon-move-in 1.2s 1s forwards;
+}
+
+@keyframes moon-move-in {
+  from {
+    opacity: 0;
+    background-position: right 150%;
+  }
+  to {
+    opacity: 1;
+    background-position: top right;
+  }
+}
+
 .container {
   width: fit-content;
   position: absolute;
   right: 0;
   top: 50%;
-  left: 50%;
+  left: 55%;
   opacity: 0;
   transform: translate(-50%, -50%);
   animation: text-fade-in 1000ms 800ms forwards;
@@ -87,58 +138,83 @@
   font-weight: normal;
   color: #fafafa;
   line-height: 32px;
+  margin-bottom: 24px;
 }
 
-button,
-button:visited {
+@media only screen and (max-width: 649px) {
+  .container {
+    right: 50%;
+    top: 10%;
+    width: 80%;
+    transform: translate(50%, 0);
+    animation: text-fade-in 1000ms 800ms forwards;
+  }
+
+  @keyframes text-fade-in {
+    from {
+      right: 0;
+    }
+    to {
+      opacity: 1;
+      right: 50%;
+    }
+  }
+
+  .silhouette {
     width: 100%;
-    max-width: 240px;
-    padding: 1rem;
-    border: 1px solid #fafafa;
-    color: #fafafa;
-    text-align: center;
-    text-transform: uppercase;
-    font-size: 1rem;
-    transition: all ease-in-out .25s;
-    cursor: pointer;
-    box-sizing: border-box;
+  }
 }
 
 button,
 button:visited {
-    background: #f300b4;
-    border: 1px solid transparent;
-    color: #fafafa;
-    font-weight: bold;
-    outline: none;
+  width: 100%;
+  max-width: 240px;
+  padding: 1rem;
+  border: 1px solid #fafafa;
+  color: #fafafa;
+  text-align: center;
+  text-transform: uppercase;
+  font-size: 1rem;
+  transition: all ease-in-out 0.25s;
+  cursor: pointer;
+  box-sizing: border-box;
+}
+
+button,
+button:visited {
+  background: #f300b4;
+  border: 1px solid transparent;
+  color: #fafafa;
+  font-weight: bold;
+  outline: none;
 }
 
 button:hover {
-    background: transparent;
-    border: 1px solid #f300b4;
-    color: #f300b4;
-    transition: all ease-in-out 0s;
-    outline: none;
+  background: transparent;
+  border: 1px solid #f300b4;
+  color: #f300b4;
+  transition: all ease-in-out 0s;
+  outline: none;
 }
 
 button:active {
-    background: transparent;
-    border: 1px solid #f300b4;
-    color: #f300b4;
-    transition: all ease-in-out 0s;
-    outline: none;
-    padding: 1.05rem 1rem calc(.95rem - 1px) calc(1rem - 1px);
-    border-top: 2px solid;
-    border-right: 2px solid;
+  background: transparent;
+  border: 1px solid #f300b4;
+  color: #f300b4;
+  transition: all ease-in-out 0s;
+  outline: none;
+  padding: 1.05rem 1rem calc(0.95rem - 1px) calc(1rem - 1px);
+  border-top: 2px solid;
+  border-right: 2px solid;
 }
 
 button:disabled {
-    opacity: .4;
-    cursor: not-allowed;
+  opacity: 0.4;
+  cursor: not-allowed;
 }
 
 button:disabled:hover {
-    background: #f300b4;
-    border: 1px solid transparent;
-    color: #fafafa;
+  background: #f300b4;
+  border: 1px solid transparent;
+  color: #fafafa;
 }
