+++ src/components/MenuHeader/index.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import Menu from "./Menu";
-import Navbar from "./Navbar";
+import Menu from "../Menu";
+import Navbar from "../Navbar";
 
 const MenuHeader = ({ bgActive }) => {
   const [isOpen, setOpen] = useState(null);
@@ -16,7 +16,7 @@ const MenuHeader = ({ bgActive }) => {
         bgActive={bgActive}
         onClickMenuButtonNav={handleClickMenuButton}
       />
-      <Menu isOpen={isOpen} />
+      <Menu isOpen={isOpen} onClickMenuButtonNav={handleClickMenuButton} />
     </>
   );
 };
