+++ src/routes/About/index.js
@@ -1,0 +1,9 @@
+const AboutPage = () => {
+  return (
+    <div>
+      <h1>This is About page</h1>
+    </div>
+  );
+};
+
+export default AboutPage;
