+++ src/routes/Contact/index.js
@@ -1,0 +1,9 @@
+const ContactPage = () => {
+  return (
+    <div>
+      <h1>This is Contact page</h1>
+    </div>
+  );
+};
+
+export default ContactPage;
