+++ src/routes/NotFound/index.js
@@ -1,0 +1,9 @@
+const NotFoundPage = () => {
+  return (
+    <div>
+      <h1>404 Not Found</h1>
+    </div>
+  );
+};
+
+export default NotFoundPage;
